server/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
beautifulsoup4==4.8.0
billiard==4.2.0
black==24.4.2
blinker==1.7.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==42.0.7
dataclasses==0.6
dataclasses-json==0.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.22.1
dvc==3.51.1
dvc-data==3.15.1
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
exceptiongroup==1.2.0
filelock==3.13.3
flake8==7.0.0
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
gigachat==0.1.22
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.48.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-images-download==2.8.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.63.0
grandalf==0.8
gto==1.7.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
hydra-core==1.3.2
idna==2.8
importlib_metadata==7.1.0
importlib_resources==6.4.0
iterative-telemetry==0.0.8
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
lxml==5.2.1
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.0.0rc6
marshmallow-enum==1.5.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.24.4
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-curand-cu12==10.3.2.106
nvidia-cusparse-cu12==12.1.0.106
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
ollama==0.2.0
oauth2client==4.1.3
ollama==0.2.0
omegaconf==2.3.0
openai==1.16.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
pathspec==0.12.1
pdoc3==0.10.0
pillow==10.3.0
pip-cache==0.0.6
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
PyDrive2==1.19.0
pyfiglet==0.8.post1
Pygments==2.17.2
pyflakes==3.2.0
pygit2==1.15.0
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-pptx==0.6.23
python-resize-image==1.1.20
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
reportlab==4.2.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scmrepo==3.3.5
selenium==4.19.0
selenium-wire==1.0.8
semver==3.0.2
serpapi==0.1.5
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqltrie==0.11.0
streamlit==1.33.0
stringcase==1.2.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==1.1.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
trio==0.25.0
trio-websocket==0.11.1
triton==2.2.0
typer==0.12.3
typing==3.7.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
voluptuous==0.14.2
watchdog==4.0.0
wcwidth==0.2.13
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.18.1
